crypto_streamlit/requirements.txt
streamlit
pandas
plotly
numpy
prophet
scikit-learn
tensorflow
newsapi-python
statsmodel
